test: split oversized merge-method test file to satisfy line-limit guardrail
test_pr_monitor_merge_methods.py grew to 1644 lines, tripping the
MAX_FIRST_PARTY_FILE_LINES (1500) maintainability guardrail. Extract the
pure classifier/helper unit tests into a new
test_pr_monitor_merge_method_classifier.py, leaving the DB-backed merge
execution integration tests in place. No behavior change.
